requirements.txt
scikit-learn
Pillow==9.5.0
numpy==1.23.5
keras==2.12.0
tensorflow==2.12.0
streamlit==1.22.0